+++ pythermalcomfort/models/heat_index_schoen.py
@@ -14,9 +14,12 @@ def heat_index_schoen(
     rh: NumericInput,
     round_output: bool = True,
 ) -> HI:
-    """Calculate the Temperature Humidity Index (THI) also known as Heat Index Schoen in accordance with the Schoen (2005) model [Schoen2005]_.
+    """Calculate the Temperature Humidity Index (THI), also known as Heat Index Schoen,
+    in accordance with the Schoen (2005) model [Schoen2005]_.
 
-    The temperature-humidity index (THI) is a simplified scale of apparent temperature, considering only dry-bulb temperature and humidity. It is another formulation of the heat index.
+    The temperature-humidity index (THI) is a simplified scale of apparent
+    temperature, considering only dry-bulb temperature and humidity. It is another
+    formulation of the heat index.
 
     Parameters
     ----------